requirements.txt
Flask
Flask-SQLAlchemy
Jinja2
pywebview
